Animation of the "guess wave" converging to the correct solution
A numerical solution to the Gross Pitaevskii Equation (GPE) with no particle-particle interactions (g=0) for the harmonic potential (k/2*(x^2)). Here we solve the GPE with the split-step algorithm. The animation shows the "guess" function (sin(x)) evolving in imaginary time and converging to the correct solution for the Harmonic Oscillator ground state: a gaussian or "bell curve."