x^2 converges to the correct GPE solution
A numerical solution to the Gross Pitaevskii Equation (GPE) with no particle-particle interactions (g=0) for the harmonic potential (k/2*(x^2)). Here we solve the GPE with the split-step algorithm. The animation shows the "guess" function (sin(x)) evolving in imaginary time and converging to the correct solution for the Harmonic Oscillator ground state: a gaussian or "bell curve." The real space spacing was 10^-3, and the imaginary time step was 10^-3 as well. The test curve (x^2) was acted upon by 800 iterations of the split step procedure, and the animation is 10 fps.