Function library for 2D GPE
A library of functions for finding the ground state of a 2D wavefunction (psi) via the split-step algorithm. Includes a function to find the energy of psi at a given time to check how badly the split-step is violating energy conservation. Has code to produce animations and static plots of psi as it evolves under the split-step algorithm.